nrf24I01p_test_successful

+++ f4-balance-bike/User/include/wireless.h
@@ -20,7 +20,7 @@ typedef struct {
     CommandType type;
     union {
         struct {
-            uint16_t speed[2];
+            int speed[2];
         } move;
         struct {
             float angle[2];
